Move get_nearest_player_name down to world struct, move func to cmd module
This function is specific to the command, we don't want to check any
other world since a target specifier only checks the world the player is
in
